move to modules structure in server
