+++ server/index.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-console */
 import 'babel-polyfill';
 import express from 'express';
 import helmet from 'helmet';
@@ -5,7 +6,7 @@ import compression from 'compression';
 import morgan from 'morgan';
 import slashes from 'connect-slashes';
 import Loadable from 'react-loadable';
-import renderMiddleware from './middlewares/renderMiddleware';
+import renderMiddleware from './render/renderMiddleware';
 
 const { PORT } = process.env;
 
@@ -21,7 +22,6 @@ app.use(renderMiddleware);
 
 Loadable.preloadAll().then(() => {
   app.listen(PORT, () => {
-    // eslint-disable-next-line
-    console.info(`pwa is running as ${__PWA_ENV__} on port ${PORT}`);
+    console.log(`pwa is running as ${__PWA_ENV__} on port ${PORT}`);
   });
 });
